feat: test attempt at optimization plugin
